countValid {statgenHTP}	R Documentation

Count valid observations per time point for a given trait

Description

Count valid observations per time point for a given trait.

Usage

countValid(TP, trait)

Arguments

TP
An object of class TP.

trait
A character string indicating the trait for which valid observations should be counted.

Value

A named numerical vector with he number of valid observations per time point .

Examples

## Create a TP object containing the data from the Phenovator.
phenoTP <- createTimePoints(dat = PhenovatorDat1,
                            experimentName = "Phenovator",
                            genotype = "Genotype",
                            timePoint = "timepoints",
                            repId = "Replicate",
                            plotId = "pos",
                            rowNum = "y", colNum = "x",
                            addCheck = TRUE,
                            checkGenotypes = c("check1", "check2",
                                               "check3", "check4"))
## Count valid observations for EffpsII per time point.
validPheno <- countValid(phenoTP, trait = "EffpsII")
head(validPheno)

[Package statgenHTP version 1.0.9.1 Index]