removeTimePoints {statgenHTP}	R Documentation

Remove time points from an object of class TP

Description

Function for removing selected time points from an object of class TP.

Usage

removeTimePoints(TP, timePoints)

Arguments

TP
An object of class TP.

timePoints
A character or numeric vector indicating the time points to be removed. When using a character string to reference a time point, the value has to be an exact match to one of the existing timePoints. When using a number it will be matched by its number ("timeNumber") in the timePoints attribute of the TP object.

Value

An object of class TP, the input with the selected time points removed.

See Also

Other functions for data preparation: as.data.frame.TP(), createTimePoints(), getTimePoints(), plot.TP(), summary.TP()

Examples

## Create a TP object containing the data from the Phenovator.
phenoTP <- createTimePoints(dat = PhenovatorDat1,
                            experimentName = "Phenovator",
                            genotype = "Genotype",
                            timePoint = "timepoints",
                            repId = "Replicate",
                            plotId = "pos",
                            rowNum = "y", colNum = "x",
                            addCheck = TRUE,
                            checkGenotypes = c("check1", "check2",
                                               "check3","check4"))
## Remove the first and last time point from the TP object.
phenoTPNew <- removeTimePoints(phenoTP,
                               timePoints = c(1, 73))

## Compare by looking at summaries.
summary(phenoTP)
summary(phenoTPNew)

[Package statgenHTP version 1.0.9.1 Index]