summary.TP {statgenHTP}	R Documentation

Summary function for TP objects

Description

Function for creating a short summary of the contents of a TP object. The summary consists of the name of the experiment, the number of time points, the first and last time point and the genotypes defined as checks.

Usage

## S3 method for class 'TP'
summary(object, ...)

Arguments

object
An object of class TP.

...
Ignored.

Value

No return value, a summary is printed.

See Also

Other functions for data preparation: as.data.frame.TP(), createTimePoints(), getTimePoints(), plot.TP(), removeTimePoints()

Examples

## Create a TP object containing the data from the Phenovator.
phenoTP <- createTimePoints(dat = PhenovatorDat1,
                            experimentName = "Phenovator",
                            genotype = "Genotype",
                            timePoint = "timepoints",
                            repId = "Replicate",
                            plotId = "pos",
                            rowNum = "y", colNum = "x",
                            addCheck = TRUE,
                            checkGenotypes = c("check1", "check2",
                                               "check3","check4"))
## Create a summary.
summary(phenoTP)

[Package statgenHTP version 1.0.9.1 Index]