summary.fitMod {statgenHTP}	R Documentation

Summary function for fitMod objects

Description

Function for creating a short summary of the contents of a TP object. The summary consists of the name of the experiment, the number of time points, the engine used to fit the models and, in case spatial models where fitted using asreml, the selected spatial model.

Usage

## S3 method for class 'fitMod'
summary(object, ...)

Arguments

object
An object of class fitMod.

...
Ignored.

Value

No return value, a summary is printed.

See Also

Other functions for spatial modeling: fitModels(), getCorrected(), getEffDims(), getGenoPred(), getHerit(), getVar(), plot.fitMod()

Examples

## Using the first example dataset (PhenovatorDat1):
## Create an object of class TP.
phenoTP <- createTimePoints(dat = PhenovatorDat1,
                            experimentName = "Phenovator",
                            genotype = "Genotype",
                            timePoint = "timepoints",
                            repId = "Replicate",
                            plotId = "pos",
                            rowNum = "y", colNum = "x",
                            addCheck = TRUE,
                            checkGenotypes = c("check1", "check2",
                                               "check3", "check4"))

## Fit a SpATS model on few time points:
modPhenoSp <- fitModels(TP = phenoTP,
                        trait = "EffpsII",
                        timePoints = c(1, 6, 36))

## Create a summary.
summary(modPhenoSp)

[Package statgenHTP version 1.0.9.1 Index]