Extract estimates from fitted splines.

Function for extracting parameter estimates from fitted splines on a specified interval.

estimateSplineParameters(
  HTPSpline,
  estimate = c("predictions", "derivatives", "derivatives2"),
  what = c("min", "max", "mean", "AUC", "p"),
  timeMin = NULL,
  timeMax = NULL,
  genotypes = NULL,
  plotIds = NULL
)

Arguments

HTPSpline
An object of class HTPSpline, the output of the fitSpline function.

estimate
The type of estimate that should be extracted, the predictions, the first derivatives or the second derivatives ("derivatives2")

what
The type of estimate that should be extracted. Either minimum ("min"), maximum ("max"), mean, area under the curve ("AUC") or a percentile. Percentiles should be given as p + percentile. E.g. for the 10th percentile specify what = "p10"

timeMin
The lower bound of the time interval from which the estimates should be extracted. If NULL the smallest time value for which the splines were fitted is used.

timeMax
The upper bound of the time interval from which the estimates should be extracted. If NULL the largest time value for which the splines were fitted is used.

genotypes
A character vector indicating the genotypes for which splines are fitted. If NULL, splines will be fitted for all genotypes.

plotIds
A character vector indicating the plotIds for which splines are fitted. If NULL, splines will be fitted for all plotIds.

See also

Examples

## Run the function to fit P-splines on a subset of genotypes.
subGeno <- c("G160", "G151")
fit.spline <- fitSpline(inDat = spatCorrectedVator,
                        trait = "EffpsII_corr",
                        genotypes = subGeno,
                        knots = 50)

## Estimate the maximum value of the trait at the beginning of the time course.
paramVator1 <- estimateSplineParameters(HTPSpline = fit.spline,
                                        estimate = "predictions",
                                        what = "max",
                                        timeMin = 1527784620,
                                        timeMax = 1528500000,
                                        genotypes = subGeno)