Extract estimates from fitted splines.

Function for extracting parameter estimates from fitted splines on a specified interval.

estimateSplineParameters(
  HTPSpline,
  estimate = c("predictions", "derivatives"),
  what = c("min", "max", "mean"),
  timeMin = NULL,
  timeMax = NULL,
  genotypes = NULL,
  plotIds = NULL
)

Arguments

HTPSpline
An object of class HTPSpline, the output of the fitSpline function.

estimate
The type of estimate that should be extracted, the predictions or the first derivatives.

what
The type of estimate that should be extracted.

timeMin
The lower bound of the time interval from which the estimates should be extracted. If NULL the smallest time value for which the splines were fitted is used.

timeMax
The upper bound of the time interval from which the estimates should be extracted. If NULL the largest time value for which the splines were fitted is used.

genotypes
A character vector indicating the genotypes for which splines are fitted. If NULL, splines will be fitted for all genotypes.

plotIds
A character vector indicating the plotIds for which splines are fitted. If NULL, splines will be fitted for all plotIds.

Examples

# Format the timepoint
spatCorrVator$timePoint <- lubridate::as_datetime(spatCorrVator$timePoint)
# Run the function to fit p-spline using the mgcv package on a subset of
# genotypes
subGeno <- c("G70","G160","G151","G179","G175","G4","G55")
fit.spline <- fitSpline(corrDat = spatCorrVator,
                        trait = "EffpsII_corr",
                        genotypes = subGeno,
                        knots = 50,
                        perMinTP = 0.8)

# Estimate the maximum value of the trait at the beginning of the time course
paramVator1 <-
     estimateSplineParameters(HTPSpline = fit.spline,
                              estimate = "predictions",
                              what = "max",
                              timeMin = 1527784620,
                              timeMax = 1528500000,
                              genotypes = subGeno)