Replace outliers for series of observations by NA

Function for replacing outliers for series of observations in the data by NA. The input can either be a data.frame, specified in dat, or the output of the fitSpline function, specified in fitSpline. Exactly one of these should be provided as input for the function.

removeSerieOut(
  dat = NULL,
  fitSpline = NULL,
  serieOut,
  trait = attr(x = serieOut, which = "trait")
)

Arguments

dat
A data.frame.

fitSpline
An object of class HTPSpline, the output of the fitSpline function.

serieOut
A data.frame with at least the column plotId with values corresponding to those in dat/fitSpline.

trait
The trait that should be replaced by NA. Can be ignored when using the output of detectSerieOut as input.

Value

Depending on the input either a data.frame or an object of class HTPSpline for which the outliers specified in serieOut are replace by NA.

See also

Examples

## Run the function to fit P-spline on a subset of genotypes.
subGenoVator <- c("G160", "G151")
fit.spline <- fitSpline(inDat = spatCorrectedVator,
                        trait = "EffpsII_corr",
                        genotypes = subGenoVator,
                        knots = 50)

## Extract the tables of predicted values and P-spline coefficients.
predDat <- fit.spline$predDat
coefDat <- fit.spline$coefDat

## The coefficients are then used to tag suspect time courses
outVator <- detectSerieOut(corrDat = spatCorrectedVator,
                           predDat = predDat,
                           coefDat = coefDat,
                           trait = "EffpsII_corr",
                           genotypes = subGenoVator,
                           thrCor = 0.9,
                           thrPca = 30)

## Replace the outliers by NA in the corrected data.
spatCorrectedVatorOut <- removeSerieOut(dat = spatCorrectedVator,
                                        serieOut = outVator)