statgenHTP: High Throughput Phenotyping (HTP) Data Analysis

Phenotypic analysis of data coming from high throughput phenotyping (HTP) platforms, including different types of outlier detection, spatial analysis, and parameter estimation. The package is being developed within the EPPN2020 project (<https://eppn2020.plant-phenotyping.eu/>). Some functions have been created to be used in conjunction with the R package 'asreml' for the 'ASReml' software, which can be obtained upon purchase from 'VSN' international (<https://www.vsni.co.uk/software/asreml>).

See also

Author

Maintainer: Bart-Jan van Rossum bart-jan.vanrossum@wur.nl (ORCID)

Authors:

Emilie J Millet emilie.millet@wur.nl (ORCID)

Maria Xose Rodriguez Alvarez (ORCID)

Diana Marcela Perez Valencia (ORCID)

Isabelle Sanchez

Nadine Hilgert

Fred van Eeuwijk (ORCID)

Martin Boer